Stop unwrapping when composing.
The choice to unwrap `self` in `composable` was originally driven by
the idea that the `composable` wrapper existed only to provide the
composition, and was therefore useless once the composition happened.

The choice to also unwrap `other` in `composable` was then done for
consistency.

Both were also slightly motivated by

* a deside for less overhead when calling a composed function, and
* a desire for cleaner easier-to-read `repr` results.

However, in principle, unwrapping is actually changing semantics in
a subtle way that's bad in edge cases (for example, I would expect
that `sacompose(..., f).functions[0] is f` is always true, and that
`f | ...` does the same thing as `composable(sacompose(..., f))`,
but if `composable` wrappers are stripped off by `|` compostion,
then this breaks down and `(f | ...).functions[0] is not f`.

And in practice, getting the unwrapping to be acceptably correct
or consistent against various edge cases is somewhere between
impossible and requiring too much coding effort to be worth it.