Switch README to Markdown format
Markdown is overwhelmingly the defacto
standard at this point, supported by
PyPI, and in this package I don't have
radical portability goals so I can use
the setuptools feature of setting the
long description content type.

reStructuredText is often treated as a
second-class citizen markup language,
has less tooling and library support,
and is currently broken on GitHub in a
way that effects this README (links
with `code` formatting).

Markdown implementations have gradually
gained reStructuredText features, and I
am not actually using/wanting any of the
ways in which reStructuredText is better
than Markdown.

The biggest downside is that Pytest can't
doctest Markdown as-is out-of-the-box: it
either needs an extra plugin like the now
seemingly unmaintained (but otherwise
almost perfect) `pytest-markdoctest`, or
the `--doctest-glob='*.md'` flag and an
extra empty line at the end of every
doctest block before the "```" end fence.